+++ test/test_large_data.py
@@ -1,13 +1,19 @@
 import gzip
 from pathlib import Path
 
-import pytest
-
-from aio_overpass import Query, Client
+from aio_overpass import Client, Query
 from aio_overpass.element import collect_elements
-from aio_overpass.pt import collect_routes, RouteQuery
+from aio_overpass.pt import RouteQuery, collect_routes
 from aio_overpass.pt_ordered import collect_ordered_routes
-from test.util import URL_INTERPRETER, mock_response, VerifyingQueryRunner, verify_element, verify_route
+from test.util import (
+    URL_INTERPRETER,
+    VerifyingQueryRunner,
+    mock_response,
+    verify_element,
+    verify_route,
+)
+
+import pytest
 
 
 @pytest.mark.asyncio